Move SignupScreen to a separate directory
